+++ packages/core/tests/css.test.ts
@@ -6,7 +6,6 @@ import {
   normalizeCssLoaderOptions,
   pluginCss,
 } from '../src/plugins/css';
-import { pluginSass } from '../src/plugins/sass';
 
 describe('normalizeCssLoaderOptions', () => {
   it('should enable exportOnlyLocals correctly', () => {
@@ -169,50 +168,6 @@ describe('plugin-css injectStyles', () => {
 
     const bundlerConfigs = await rsbuild.initConfigs();
 
-    expect(bundlerConfigs[0]).toMatchSnapshot();
-  });
-});
-
-describe('plugin-sass', () => {
-  it('should add sass-loader', async () => {
-    const rsbuild = await createStubRsbuild({
-      plugins: [pluginSass()],
-      rsbuildConfig: {
-        tools: {},
-      },
-    });
-
-    const bundlerConfigs = await rsbuild.initConfigs();
-    expect(bundlerConfigs[0]).toMatchSnapshot();
-  });
-
-  it('should add sass-loader and css-loader when injectStyles', async () => {
-    const rsbuild = await createStubRsbuild({
-      plugins: [pluginSass()],
-      rsbuildConfig: {
-        output: {
-          injectStyles: true,
-        },
-      },
-    });
-
-    const bundlerConfigs = await rsbuild.initConfigs();
-    expect(bundlerConfigs[0]).toMatchSnapshot();
-  });
-
-  it('should add sass-loader with excludes', async () => {
-    const rsbuild = await createStubRsbuild({
-      plugins: [pluginSass()],
-      rsbuildConfig: {
-        tools: {
-          sass(_config, { addExcludes }) {
-            addExcludes(/node_modules/);
-          },
-        },
-      },
-    });
-
-    const bundlerConfigs = await rsbuild.initConfigs();
     expect(bundlerConfigs[0]).toMatchSnapshot();
   });
 });
