+++ packages/plugin-sass/modern.config.ts
@@ -1,0 +1,3 @@
+import { configForDualPackage } from '../../scripts/modern.base.config';
+
+export default configForDualPackage;
