+++ packages/plugin-sass/tests/index.test.ts
@@ -1,0 +1,46 @@
+import { createStubRsbuild } from '@scripts/test-helper';
+import { pluginSass } from '../src';
+
+describe('plugin-sass', () => {
+  it('should add sass-loader', async () => {
+    const rsbuild = await createStubRsbuild({
+      plugins: [pluginSass()],
+      rsbuildConfig: {
+        tools: {},
+      },
+    });
+
+    const bundlerConfigs = await rsbuild.initConfigs();
+    expect(bundlerConfigs[0]).toMatchSnapshot();
+  });
+
+  it('should add sass-loader and css-loader when injectStyles', async () => {
+    const rsbuild = await createStubRsbuild({
+      plugins: [pluginSass()],
+      rsbuildConfig: {
+        output: {
+          injectStyles: true,
+        },
+      },
+    });
+
+    const bundlerConfigs = await rsbuild.initConfigs();
+    expect(bundlerConfigs[0]).toMatchSnapshot();
+  });
+
+  it('should add sass-loader with excludes', async () => {
+    const rsbuild = await createStubRsbuild({
+      plugins: [pluginSass()],
+      rsbuildConfig: {
+        tools: {
+          sass(_config, { addExcludes }) {
+            addExcludes(/node_modules/);
+          },
+        },
+      },
+    });
+
+    const bundlerConfigs = await rsbuild.initConfigs();
+    expect(bundlerConfigs[0]).toMatchSnapshot();
+  });
+});
