+++ packages/shared/package.json
@@ -102,7 +102,6 @@
     "@types/connect": "3.4.38",
     "@types/fs-extra": "^11.0.4",
     "@types/node": "18.x",
-    "@types/sass-loader": "8.0.8",
     "@types/webpack-bundle-analyzer": "4.7.0",
     "autoprefixer": "10.4.19",
     "browserslist": "4.23.0",
@@ -119,8 +118,6 @@
     "picocolors": "1.0.0",
     "prebundle": "1.1.0",
     "rslog": "^1.2.2",
-    "sass": "^1.77.1",
-    "sass-loader": "14.2.1",
     "semver": "^7.6.2",
     "terser": "5.31.0",
     "typescript": "^5.4.2",
